content-security-policy/script-src/script-src-strict_dynamic_double_policy_report_only.html

<!DOCTYPE HTML>
<html>

<head>
    <title>A separate Report-Only policy does not influence `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>
    <script src='../support/checkSecurityPolicyViolations.sub.js' nonce='dummy'></script>
    <!-- CSP served:
    1) Content-Security-Policy: script-src 'strict-dynamic' 'nonce-dummy'
    2) Content-Security-Policy-Report-Only: script-src 'none'
    -->
</head>

<body>
    <h1>A separate Report-Only policy does not influence `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('script');
            e.src = 'simpleSourcedScript.js';
            e.onload = t.step_func_done(function() {
                assert_true(simpleSourcedScriptRan);
            });
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
        }, 'Script injected via `appendChild` is allowed with `strict-dynamic` + Report-Only `script-src \'none\'` policy.');
    </script>

    <script nonce='dummy'>
        async_test(function(t) {
            setTimeout(t.step_func_done(function() {
                // Check that every violation comes from the Report-Only policy.
                assert_original_policies(["script-src 'none'"]);
            }), 1000);
        }, 'All the CSP violation reports have been fired.');
    </script>

</body>

</html>
