LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_double_policy_report_only-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
CONSOLE MESSAGE: The Content Security Policy 'script-src 'none'' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?appendChild-reportOnly because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?appendChild-reportOnly because it does not appear in the script-src directive of the Content Security Policy.
A separate Report-Only policy does not influence `strict-dynamic` in the script-src directive.


FAIL Script injected via `appendChild` is allowed with `strict-dynamic` + Report-Only `script-src 'none'` policy. assert_unreached: Error should not be triggered. Reached unreachable code
